Update README.md
How to Create and Configure a Swap File on a BTT-CB1

This tutorial will guide you through the process of creating and configuring a swap file on a BTT-CB1. Increasing swap space can help manage memory usage and improve system stability for memory-intensive tasks.

Prerequisites

	•	A BTT-CB1 running a Debian-based distribution.
	•	Basic knowledge of using the terminal.

Introduction

Swap space is used to extend the amount of memory available to applications by allowing the operating system to offload data from RAM to a designated area on disk when physical memory is full. This can prevent out-of-memory errors and improve system stability, especially when running memory-intensive applications.

This tutorial will show you how to create and configure a swap file on a BTT-CB1.

Step-by-Step Guide